# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Education Mini Program API
  version: 0.1.0
  description: >-
    Shared API contract for the first-phase education mini program. Successful
    responses use a data envelope and failures preserve their HTTP status.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Environment-relative service root
...
security:
  - bearerAuth: []
paths:
...
  /teachers/me/withdrawals:
    get:
      operationId: listTeacherWithdrawals
      summary: List withdrawal requests for the authenticated teacher
      tags:
        - Teacher
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/WithdrawalStatus'
      responses:
        '200':
          description: Paginated teacher withdrawal requests.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherWithdrawalsPageEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createTeacherWithdrawal
      summary: Reserve available earnings in a new withdrawal request
      tags:
        - Teacher
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeacherWithdrawalRequest'
      responses:
        '200':
          description: Created or idempotently replayed withdrawal request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherWithdrawalEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /teachers/me/withdrawals/{withdrawalId}/cancel:
    post:
      operationId: cancelTeacherWithdrawal
      summary: Cancel the authenticated teacher's submitted withdrawal
      tags:
        - Teacher
      parameters:
        - $ref: '#/components/parameters/WithdrawalId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedActionRequest'
      responses:
        '200':
          description: Cancelled or idempotently replayed withdrawal request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherWithdrawalEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /management/teacher-withdrawal-policies:
    get:
      operationId: listTeacherWithdrawalPolicies
      summary: List versioned teacher withdrawal policies
      tags:
        - Management
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ConfigurationStatus'
      responses:
        '200':
          description: Paginated withdrawal policies.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherWithdrawalPoliciesPageEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createTeacherWithdrawalPolicy
      summary: Create a draft teacher withdrawal-policy version
      tags:
        - Management
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeacherWithdrawalPolicyRequest'
      responses:
        '200':
          description: Created or idempotently replayed withdrawal policy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherWithdrawalPolicyEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /management/teacher-withdrawal-policies/{policyId}/activate:
    post:
      operationId: activateTeacherWithdrawalPolicy
      summary: Activate a draft withdrawal-policy version
      tags:
        - Management
      parameters:
        - $ref: '#/components/parameters/WithdrawalPolicyId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedActionRequest'
      responses:
        '200':
          description: Activated withdrawal-policy version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherWithdrawalPolicyEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /management/teacher-withdrawal-policies/{policyId}/retire:
    post:
      operationId: retireTeacherWithdrawalPolicy
      summary: Retire an active withdrawal-policy version
      tags:
        - Management
      parameters:
        - $ref: '#/components/parameters/WithdrawalPolicyId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetireConfigurationRequest'
      responses:
        '200':
          description: Retired withdrawal-policy version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherWithdrawalPolicyEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /management/withdrawals:
    get:
      operationId: listManagedTeacherWithdrawals
      summary: List teacher withdrawal requests for review and manual payout
      tags:
        - Management
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: campusId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/WithdrawalStatus'
      responses:
        '200':
          description: Paginated teacher withdrawal requests.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherWithdrawalsPageEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /management/withdrawals/{withdrawalId}/approve:
    post:
      operationId: approveTeacherWithdrawal
      summary: Approve a submitted teacher withdrawal
      tags:
        - Management
      parameters:
        - $ref: '#/components/parameters/WithdrawalId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedActionRequest'
      responses:
        '200':
          description: Approved teacher withdrawal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherWithdrawalEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /management/withdrawals/{withdrawalId}/reject:
    post:
      operationId: rejectTeacherWithdrawal
      summary: Reject a submitted teacher withdrawal and release allocations
      tags:
        - Management
      parameters:
        - $ref: '#/components/parameters/WithdrawalId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonedActionRequest'
      responses:
        '200':
          description: Rejected teacher withdrawal.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherWithdrawalEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /management/withdrawals/{withdrawalId}/mark-paying:
    post:
      operationId: markTeacherWithdrawalPaying
      summary: Record that manual payout processing has started
      tags:
        - Management
      parameters:
        - $ref: '#/components/parameters/WithdrawalId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedActionRequest'
      responses:
        '200':
          description: Withdrawal marked as paying.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherWithdrawalEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /management/withdrawals/{withdrawalId}/mark-paid:
    post:
      operationId: markTeacherWithdrawalPaid
      summary: Record a completed manual payout with proof
      tags:
        - Management
      parameters:
        - $ref: '#/components/parameters/WithdrawalId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkWithdrawalPaidRequest'
      responses:
        '200':
          description: Withdrawal marked as paid after manual proof validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherWithdrawalEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /management/withdrawals/{withdrawalId}/mark-failed:
    post:
      operationId: markTeacherWithdrawalFailed
      summary: Record a failed manual payout and release allocations
      tags:
        - Management
      parameters:
        - $ref: '#/components/parameters/WithdrawalId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonedActionRequest'
      responses:
        '200':
          description: Withdrawal marked as failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeacherWithdrawalEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by the authentication module.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Caller-generated key used by mutation endpoints to make retries safe.
        Keys contain 8-128 letters, numbers, dots, underscores, colons, or
        hyphens.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]{8,128}$'
      example: lesson-confirm-20260827-0001
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    WithdrawalId:
      name: withdrawalId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    WithdrawalPolicyId:
      name: policyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    RequestId:
      description: Correlation identifier matching the response body requestId.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$'
...
  responses:
...
    BadRequest:
      description: Request validation failed.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Authentication is required or the session is invalid.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: The authenticated user lacks the required role or data scope.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: The requested state transition conflicts with current state.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ResourceNotFound:
      description: The requested resource does not exist.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            missingResource:
              value:
                code: RESOURCE_NOT_FOUND
                message: Cannot GET /resource
                requestId: 64ea9657-41f1-4baf-a3fb-d62e39054016
                details: {}
    InternalServerError:
      description: An unexpected server error occurred.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            internalError:
              value:
                code: INTERNAL_SERVER_ERROR
                message: Internal server error
                requestId: 64ea9657-41f1-4baf-a3fb-d62e39054016
                details: {}
  schemas:
...
    SuccessEnvelope:
      type: object
      description: Common wrapper for every successful response.
      required:
        - data
        - requestId
      properties:
        data:
          description: Endpoint-specific response payload.
        requestId:
          type: string
          minLength: 1
          maxLength: 128
          description: Correlation identifier returned in X-Request-Id.
      additionalProperties: false
    PaginatedSuccessEnvelope:
      type: object
      description: Common wrapper for paginated list responses.
      required:
        - data
        - meta
        - requestId
      properties:
        data:
          type: array
          items: {}
        meta:
          $ref: '#/components/schemas/PaginationMeta'
        requestId:
          type: string
          minLength: 1
          maxLength: 128
          description: Correlation identifier returned in X-Request-Id.
      additionalProperties: false
...
    ErrorEnvelope:
      type: object
      description: Common response body for failures; the HTTP status is preserved.
      required:
        - code
        - message
        - requestId
        - details
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          minLength: 1
        requestId:
          type: string
          minLength: 1
          maxLength: 128
        details:
          type: object
          additionalProperties: true
      additionalProperties: false
    ErrorCode:
      type: string
      enum:
        - BAD_REQUEST
        - VALIDATION_FAILED
        - INVALID_IDEMPOTENCY_KEY
        - UNAUTHORIZED
        - FORBIDDEN
        - RESOURCE_NOT_FOUND
        - METHOD_NOT_ALLOWED
        - CONFLICT
        - STAFF_PHONE_ALREADY_EXISTS
        - STAFF_ACCOUNT_UNAVAILABLE
        - STAFF_PHONE_BINDING_REQUIRED
        - STAFF_ACCOUNT_VERSION_CONFLICT
        - WECHAT_IDENTITY_CONFLICT
        - WECHAT_PROVIDER_RESPONSE_INVALID
        - LESSON_NOT_ASSIGNED
        - LESSON_ALREADY_COMPLETED
        - LESSON_VERSION_CONFLICT
        - LESSON_REVERSAL_WINDOW_EXPIRED
        - ATTENDANCE_INCOMPLETE
        - INSUFFICIENT_LESSON_BALANCE
        - FEEDBACK_NOT_ALLOWED
        - FEEDBACK_IMAGE_INVALID
        - FEEDBACK_IMAGE_CONFLICT
        - GROUP_CAMPAIGN_POSTER_INVALID
        - CUSTOMER_SERVICE_QR_INVALID
        - IDEMPOTENCY_CONFLICT
        - EARNING_RULE_NOT_FOUND
        - EARNING_RULE_OVERLAP
        - EARNING_NOT_REVIEWABLE
        - EARNING_ALREADY_REVIEWED
        - EARNING_REVERSED
        - WITHDRAWAL_AMOUNT_INVALID
        - WITHDRAWAL_BALANCE_INSUFFICIENT
        - WITHDRAWAL_FREQUENCY_EXCEEDED
        - WITHDRAWAL_STATUS_CONFLICT
        - WITHDRAWAL_SELF_REVIEW_FORBIDDEN
        - PAYOUT_PROOF_REQUIRED
        - PAYOUT_PROOF_INVALID
        - TOO_MANY_REQUESTS
        - INTERNAL_SERVER_ERROR
    PaginationMeta:
      type: object
      description: Metadata returned by paginated list endpoints.
      required:
        - page
        - pageSize
        - total
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
      additionalProperties: false
...
    WithdrawalStatus:
      type: string
      enum:
        - SUBMITTED
        - APPROVED
        - PAYING
        - PAID
        - CANCELLED
        - REJECTED
        - FAILED
    ConfigurationStatus:
      type: string
      enum:
        - DRAFT
        - ACTIVE
        - RETIRED
...
    TeacherWithdrawalView:
      type: object
      required:
        - id
        - requestNo
        - campusId
        - teacherId
        - teacherName
        - amountFen
        - status
        - requestedAt
        - reviewedAt
        - paidAt
        - rejectionReason
        - failureReason
        - payoutReference
        - payoutProofFileId
        - version
      properties:
        id:
          type: string
          format: uuid
        requestNo:
          type: string
          minLength: 1
          maxLength: 64
        campusId:
          type: string
          format: uuid
        teacherId:
          type: string
          format: uuid
        teacherName:
          type: string
          minLength: 1
          maxLength: 100
        amountFen:
          type: integer
          minimum: 1
        status:
          $ref: '#/components/schemas/WithdrawalStatus'
        requestedAt:
          type: string
          format: date-time
        reviewedAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        paidAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        rejectionReason:
          oneOf:
            - type: string
              minLength: 1
              maxLength: 500
            - type: 'null'
        failureReason:
          oneOf:
            - type: string
              minLength: 1
              maxLength: 500
            - type: 'null'
        payoutReference:
          oneOf:
            - type: string
              minLength: 1
              maxLength: 128
            - type: 'null'
        payoutProofFileId:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
        version:
          type: integer
          minimum: 1
      additionalProperties: false
    TeacherWithdrawalPolicyView:
      type: object
      required:
        - id
        - minimumAmountFen
        - dailyRequestLimit
        - version
        - status
        - effectiveFrom
        - effectiveTo
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        minimumAmountFen:
          type: integer
          minimum: 1
        dailyRequestLimit:
          type: integer
          minimum: 1
        version:
          type: integer
          minimum: 1
        status:
          $ref: '#/components/schemas/ConfigurationStatus'
        effectiveFrom:
          type: string
          format: date-time
        effectiveTo:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        createdAt:
          type: string
          format: date-time
      additionalProperties: false
    CreateTeacherWithdrawalRequest:
      type: object
      required:
        - amountFen
      properties:
        amountFen:
          type: integer
          minimum: 1
      additionalProperties: false
...
    CreateTeacherWithdrawalPolicyRequest:
      type: object
      required:
        - minimumAmountFen
        - dailyRequestLimit
        - effectiveFrom
      properties:
        minimumAmountFen:
          type: integer
          minimum: 1
        dailyRequestLimit:
          type: integer
          minimum: 1
        effectiveFrom:
          type: string
          format: date-time
      additionalProperties: false
...
    VersionedActionRequest:
      type: object
      required:
        - expectedVersion
      properties:
        expectedVersion:
          type: integer
          minimum: 1
      additionalProperties: false
    RetireConfigurationRequest:
      type: object
      required:
        - expectedVersion
        - effectiveTo
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        effectiveTo:
          type: string
          format: date-time
      additionalProperties: false
    ReasonedActionRequest:
      type: object
      required:
        - expectedVersion
        - reason
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        reason:
          type: string
          minLength: 1
          maxLength: 500
      additionalProperties: false
    MarkWithdrawalPaidRequest:
      type: object
      required:
        - expectedVersion
        - payoutReference
        - payoutProofFileId
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        payoutReference:
          type: string
          minLength: 1
          maxLength: 128
        payoutProofFileId:
          type: string
          format: uuid
      additionalProperties: false
...
    TeacherWithdrawalEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TeacherWithdrawalView'
    TeacherWithdrawalsPageEnvelope:
      allOf:
        - $ref: '#/components/schemas/PaginatedSuccessEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/TeacherWithdrawalView'
...
    TeacherWithdrawalPolicyEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TeacherWithdrawalPolicyView'
    TeacherWithdrawalPoliciesPageEnvelope:
      allOf:
        - $ref: '#/components/schemas/PaginatedSuccessEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/TeacherWithdrawalPolicyView'
...